+++ package/PinchZoomSliderModal/package.json
@@ -2,7 +2,7 @@
   "name": "@dooboo-ui/pinch-zoom-slider-modal",
   "main": "package/index.js",
   "types": "package/index.d.ts",
-  "version": "0.0.5",
+  "version": "0.0.1",
   "license": "MIT",
   "postinstall": "tsc",
   "scripts": {
