+++ package/theme/package.json
@@ -2,7 +2,7 @@
   "name": "@dooboo-ui/theme",
   "main": "package/index.js",
   "types": "package/index.d.ts",
-  "version": "0.1.7",
+  "version": "0.0.1",
   "license": "MIT",
   "scripts": {
     "build": "rm -rf ./package && tsc"
