+++ storybook/stories/index.js
@@ -1,13 +1,13 @@
-import './Accordian.stories';
-import './Button.stories';
-import './ButtonGroup.stories';
-import './EditText.stories';
-import './LoadingIndicator.stories';
-import './RadioButton.stories';
-import './SearchInput.stories';
-import './Snackbar.stories';
-import './SwitchToggle.stories';
-import './TinderCard.stories';
-import './Slider.stories';
-import './PinchZoomSliderModal.stories';
-import './Rating.stories';
+import './shared/Accordian.stories';
+import './shared/Button.stories';
+import './shared/ButtonGroup.stories';
+import './shared/EditText.stories';
+import './shared/LoadingIndicator.stories';
+import './shared/RadioButton.stories';
+import './shared/SearchInput.stories';
+import './shared/SwitchToggle.stories';
+import './shared/Slider.stories';
+import './shared/Rating.stories';
+import './packages/Snackbar.stories';
+import './packages/TinderCard.stories';
+import './packages/PinchZoomSliderModal.stories';
