+++ storybook/stories/shared/Button.stories.tsx
@@ -1,9 +1,9 @@
-import { IC_FACEBOOK, IC_GOOGLE } from '../../src/Icons';
+import { IC_FACEBOOK, IC_GOOGLE } from '../../../src/Icons';
 import { Image, View } from 'react-native';
 import React, { useState } from 'react';
 
-import Button from '../../src/components/shared/Button';
-import { ContainerDeco } from '../decorators';
+import Button from '../../../src/components/shared/Button';
+import { ContainerDeco } from '../../decorators';
 import { action } from '@storybook/addon-actions';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
