+++ storybook/stories/shared/Rating.stories.tsx
@@ -1,11 +1,11 @@
-import { ContainerDeco } from '../decorators';
-import Rating from '../../src/components/shared/Rating';
+// eslint-disable-next-line sort-imports
+import { boolean, number } from '@storybook/addon-knobs';
+
+import { ContainerDeco } from '../../decorators';
+import Rating from '../../../src/components/shared/Rating';
 import React from 'react';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
-
-// eslint-disable-next-line sort-imports
-import { boolean, number } from '@storybook/addon-knobs';
 
 storiesOf('Rating', module)
   .addDecorator(ContainerDeco)
