+++ storybook/stories/shared/SearchInput.stories.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { number, text } from '@storybook/addon-knobs';
 
-import { IC_MAGNIFIER } from '../../src/Icons';
-import SearchInput from '../../src/components/shared/SearchInput';
+import { IC_MAGNIFIER } from '../../../src/Icons';
+import SearchInput from '../../../src/components/shared/SearchInput';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
